Added BM3D denoising filter
Added Visual Studio property sheets for FFTW3 libraries
Added fftw3_helper.hpp with template struct fftwh<>, as a C++ interface for FFTW3

Type.h: DType changed from uint32 to sint32
Type.h: Added template struct _IsSInt, _IsUInt, _IsInt and _IsFloat for compile-time type check
Type.h: Added template functions TypeMin and TypeMax returning the min and max value for specific numeric type

Specification.h: enum ColorMatrix::Average renamed to ColorMatrix::OPP (opponent colorspace)
Specification.h: Added template functions ColorMatrix_RGB2YUV_Parameter and ColorMatrix_YUV2RGB_Parameter for YUV<->RGB conversion matrix coefficients

Image_Type.h: Added member functions SetQuantRange, SetChromaPlacement, SetColorPrim and SetColorMatrix to class Frame
Image_Type.hpp: Cosmetics

LUT.h: LUT<>::LevelType changed from uint32 to ptrdiff_t

Conversion.hpp: Added template functions MatrixConvert_RGB2YUV, MatrixConvert_YUV2RGB and MatrixConvert for YUV<->RGB conversion

Block.h: Some cosmetics to template class Block<>, float_type renamed to dist_type, added protected function SearchBoundary, fixed boundary calculation in BlockMatching and BlockMatchingMulti
Block.h: Template class BlockGroup<>, "BlockCount" renamed to "GroupSize", added member functions To, AddTo, CountTo

ImageIO.cpp: Cosmetics

NLMeans: Option "Overlap" deleted, replaced with option "BlockStep"
NLMeans: Cosmetics

BM3D.h: Added filter class BM3D_Basic, BM3D_Final and BM3D
BM3D.cpp: Added implementaion of BM3D denoise algorithm

+++ include/LUT.h
@@ -21,7 +21,7 @@ public:
     typedef pointer iterator;
     typedef const_pointer const_iterator;
 
-    typedef uint32 LevelType;
+    typedef ptrdiff_t LevelType;
 
 private:
     LevelType Levels_ = 0;
